broker updated

+++ fixme-broker/src/main/java/com/wethinkcode/fixme/broker/model/BrokerInstruments.java
@@ -4,11 +4,11 @@ import lombok.Getter;
 
 public class BrokerInstruments {
     @Getter
-    private Integer usdInstrument = 0;
+    private Double usdInstrument = 0;
     @Getter
     private Float usdPrice;
     @Getter
-    private Integer zarInstrument = 0;
+    private Double zarInstrument = 0;
     @Getter
     private Float zarPrice;
 
@@ -16,7 +16,7 @@ public class BrokerInstruments {
 
     }
 
-    public void AddInstrument(Integer quantity, Float price, String type){
+    public void AddInstrument(Float quantity, Float price, String type){
         if(type == "zar") {
             this.zarInstrument = this.zarInstrument + quantity;
             this.zarPrice = price;
